misc: cleanup
